requirements.txt
selenium
undetected-chromedriver
requests
python-docx
openpyxl
lxml
chardet
bidi
python-dotenv
